# Performance improvement, code refactoring, and automatically correcting blendshapes orders