--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,3 @@
+Compiled
+Compiled/*
+Plugin/obj