--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 Compiled
 Compiled/*
 Plugin/obj
+.vs