--- a/Backup/Project.sln
+++ b/Backup/Project.sln
@@ -19,7 +19,4 @@ Global
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4022B487-C475-46E0-9253-DDA288BC7453}
-	EndGlobalSection
 EndGlobal