--- a/Plugin/UniversalAssetLoader.csproj
+++ b/Plugin/UniversalAssetLoader.csproj
@@ -561,5 +561,9 @@
       <HintPath>..\Dependencies\UnityEngine.UI.dll</HintPath>  
       <Private>false</Private>  
     </Reference> 
+    <Reference Include="Newtonsoft.Json">  
+      <HintPath>..\Dependencies\Newtonsoft.Json.dll</HintPath>  
+      <Private>false</Private>
+    </Reference>
   </ItemGroup> 
 </Project>