… Assets/blendshape_orders.json
… Dependencies/Newtonsoft.Json.dll
… Dependencies/Newtonsoft.Json.pdb
… Dependencies/Newtonsoft.Json.xml
… Plugin/Plugin.cs