### [[optimization] Speed up the mesh loading process at startup](https://github.com/aryrk/MiSide-UniversalAssetLoader/commit/b5563fbc4295f8f67ce5e9d4d9028f4260abc75b)
When there are many meshes in Asset folder, the load process becomes quite noticeable. I changed the mesh load process to parallel with a maximum thread of `Number of CPU cores - 1`, but of course the minimum thread is 1.

The results are not that drastic but they are quite noticeable.
Originally it takes 2080ms to load all the mesh on startup, now it only takes 660ms (of course system specifications will affect)

Before:
![Screenshot (64)](https://github.com/user-attachments/assets/c356d87a-6ba5-4574-a27a-5bf90961024c)

After:
![Screenshot (68)](https://github.com/user-attachments/assets/3b71d09f-38a0-49f7-b805-421b65a561eb)

I wish this can be applied on loading ogg and texture files, but these processes need to be running on main thread :(


### [[refactoring] Improve performance, completely rewrite all necessary condition in PatchMita to switch case](https://github.com/aryrk/MiSide-UniversalAssetLoader/commit/001e0b257ddc3c0a0b2c762bb6acb5293fade154)
Seriously? It's full of if else conditions, I changed it to switch case condition.
Bro who uses if else for conditions like this 💀
I'm talking about `public static void PatchMita` in `Plugin.cs`

The result are not noticeable, but you get it,
Before, it takes 609ms
![Screenshot (71)](https://github.com/user-attachments/assets/f9e4d094-bdea-4e64-bd44-9bc9196263c1)

After, it's only takes 571ms
![Screenshot (73)](https://github.com/user-attachments/assets/4661cda6-a07e-4499-bca7-8ebd4ee09799)


### [[feat] Automatically correcting the BlendShapes orders](https://github.com/aryrk/MiSide-UniversalAssetLoader/commit/34470f033aca152c3f73d4f2ca2c013c3c7e8bc2)
As the name suggests, this will fix the order of the blendshape because the algorithm replaces according to the index instead of the shapekeys name.
The order is in the file `Assets/blendshape_orders.json`
Unknown ShapeKeys will be discarded, and if there are not enough, a dummy blendshape will be created.

`blendshape_orders.json` can be used for any other characters, just make sure to change the parameter accordingly

![Screenshot (77)](https://github.com/user-attachments/assets/b0437d5c-55f2-464f-b22b-9b8d5d4c7126)

The only downside is that the performance is not that good as before, now it takes 681ms to load all shapekeys, but still fair enough


### #**NOTE**
This needs proper testing because I'm still using the old AssimpNet.