@Rist8 2025-01-09
Tested with new version of AssimpNet.dll and it seems to be fine on my device.